# backend/swagger/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "TEAM WORKS API",
    "description": "팀 단위 캘린더 기반 일정관리 + 채팅 애플리케이션 API. API Route 소스 위치: backend/app/api/. v1.7.0 — Postits/Notices/Projects/ProjectSchedules/SubSchedules/Me/Refresh paths 보강",
    "version": "1.7.1",
    "x-source-directory": "backend/app/api"
  },
  "servers": [
    {
      "url": "/api",
      "description": "Base URL (Next.js API Routes — backend/app/api/)"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string",
            "description": "에러 메시지"
          }
        }
      },
...
      "BoardAttachment": {
        "type": "object",
        "required": [
          "id",
          "postId",
          "originalName",
          "mimeType",
          "sizeBytes",
          "uploadedAt"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "example": "c4d5e6f7-a8b9-0123-cdef-ab4567890123"
          },
          "postId": {
            "type": "string",
            "format": "uuid",
            "example": "b3c4d5e6-f7a8-9012-bcde-fa3456789012"
          },
          "originalName": {
            "type": "string",
            "description": "업로드 시 원본 파일명",
            "example": "project-plan.pdf"
          },
          "mimeType": {
            "type": "string",
            "description": "검증된 MIME 타입",
            "example": "application/pdf"
          },
          "sizeBytes": {
            "type": "integer",
            "description": "파일 크기 (바이트)",
            "example": 204800
          },
          "uploadedAt": {
            "type": "string",
            "format": "date-time",
            "example": "2026-04-20T01:00:00.000Z"
          }
        }
      },
      "BoardPost": {
        "type": "object",
        "required": [
          "id",
          "teamId",
          "authorId",
          "authorName",
          "title",
          "content",
          "createdAt",
          "updatedAt",
          "attachments"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "example": "b3c4d5e6-f7a8-9012-bcde-fa3456789012"
          },
          "teamId": {
            "type": "string",
            "format": "uuid",
            "example": "b1c2d3e4-f5a6-7890-bcde-fa1234567890"
          },
          "projectId": {
            "type": "string",
            "format": "uuid",
            "nullable": true,
            "description": "프로젝트 자료실이면 해당 UUID, 팀 일자별이면 null",
            "example": null
          },
          "authorId": {
            "type": "string",
            "format": "uuid",
            "example": "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
          },
          "authorName": {
            "type": "string",
            "description": "작성자 표시 이름",
            "example": "홍길동"
          },
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200,
            "example": "프로젝트 계획서"
          },
          "content": {
            "type": "string",
            "maxLength": 20000,
            "example": "프로젝트 초기 계획서를 첨부합니다."
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2026-04-20T01:00:00.000Z"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "example": "2026-04-20T01:00:00.000Z"
          },
          "attachments": {
            "type": "array",
            "description": "첨부파일 메타데이터 목록 (없으면 빈 배열)",
            "items": {
              "$ref": "#/components/schemas/BoardAttachment"
            }
          }
        }
      },
...
      "UpdatePostRequest": {
        "type": "object",
        "description": "자료실 글 수정 요청 (multipart/form-data). 전달된 필드만 갱신.",
        "properties": {
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200,
            "description": "새 제목",
            "example": "프로젝트 계획서 (수정)"
          },
          "content": {
            "type": "string",
            "maxLength": 20000,
            "description": "새 본문",
            "example": "수정된 계획서를 첨부합니다."
          },
          "file": {
            "type": "string",
            "format": "binary",
            "description": "첨부파일 교체. 전달 시 기존 첨부 unlink 후 신규 저장"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/teams/{teamId}/board/{postId}": {
      "get": {
        "tags": [
          "Board"
        ],
        "summary": "자료실 글 상세 조회",
        "description": "자료실 게시글 상세를 조회합니다. 첨부파일 메타데이터를 포함합니다.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "teamId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "소속 팀 UUID"
          },
          {
            "name": "postId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "조회할 게시글 UUID"
          }
        ],
        "responses": {
          "200": {
            "description": "게시글 상세 조회 성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BoardPost"
                }
              }
            }
          },
          "401": {
            "description": "인증 필요",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "인증이 필요합니다."
                }
              }
            }
          },
          "403": {
            "description": "접근 권한 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "해당 팀에 접근 권한이 없습니다."
                }
              }
            }
          },
          "404": {
            "description": "게시글 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "글을 찾을 수 없습니다."
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Board"
        ],
        "summary": "자료실 글 수정",
        "description": "자료실 게시글을 수정합니다. 작성자 본인만 수정 가능. file 전달 시 기존 첨부 unlink 후 신규 저장.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "teamId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "소속 팀 UUID"
          },
          {
            "name": "postId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "수정할 게시글 UUID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePostRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "게시글 수정 성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BoardPost"
                }
              }
            }
          },
          "400": {
            "description": "잘못된 요청",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "empty_title": {
                    "value": {
                      "error": "제목은 비울 수 없습니다."
                    }
                  },
                  "title_too_long": {
                    "value": {
                      "error": "제목은 최대 200자까지 입력 가능합니다."
                    }
                  },
                  "content_too_long": {
                    "value": {
                      "error": "본문은 최대 20000자까지 입력 가능합니다."
                    }
                  },
                  "invalid_mime": {
                    "value": {
                      "error": "허용되지 않는 파일 형식입니다."
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "인증 필요",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "인증이 필요합니다."
                }
              }
            }
          },
          "403": {
            "description": "권한 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "no_team_access": {
                    "value": {
                      "error": "해당 팀에 접근 권한이 없습니다."
                    }
                  },
                  "not_author": {
                    "value": {
                      "error": "본인이 작성한 글만 수정할 수 있습니다."
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "게시글 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "글을 찾을 수 없습니다."
                }
              }
            }
          },
          "413": {
            "description": "요청 본문 크기 초과",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "요청 본문이 너무 큽니다."
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Board"
        ],
        "summary": "자료실 글 삭제",
        "description": "자료실 게시글을 삭제합니다. 작성자 본인만 삭제 가능. 첨부파일도 함께 unlink됩니다.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "teamId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "소속 팀 UUID"
          },
          {
            "name": "postId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "삭제할 게시글 UUID"
          }
        ],
        "responses": {
          "200": {
            "description": "게시글 삭제 성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "ok"
                  ],
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "example": true
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "인증 필요",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "인증이 필요합니다."
                }
              }
            }
          },
          "403": {
            "description": "권한 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "no_team_access": {
                    "value": {
                      "error": "해당 팀에 접근 권한이 없습니다."
                    }
                  },
                  "not_author": {
                    "value": {
                      "error": "본인이 작성한 글만 삭제할 수 있습니다."
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "게시글 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "글을 찾을 수 없습니다."
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}